Use garage-check at the end of garage-deploy
Moved garage-check's logic in a small library

Added check.cc and check.h

Deleted the vim comment
